+++ src/main/java/lermitage/intellij/extra/icons/ExtraIconProvider.java
@@ -343,6 +343,8 @@ public class ExtraIconProvider extends BaseIconProvider implements DumbAware {
                 .regex(".*\\.env(\\.[a-zA-Z0-9]+)*"),
             ofFile("dotenv_alt2", "/extra-icons/env_alt2.svg", "Dotenv: .*\\.env(\\.[a-zA-Z0-9]+)* (alternative 2)")
                 .regex(".*\\.env(\\.[a-zA-Z0-9]+)*"),
+            ofFile("eslint", "/extra-icons/eslint.svg", "Eslint (regex): .*\\.eslintrc(\\.[cm]?js(on)?|\\.y(a)?ml)?$")
+                .regex(".*\\.eslintrc(\\.[cm]?js(on)?|\\.y(a)?ml)?$"),
             ofFile("tsconfig", "/extra-icons/tsconfig.svg", "TSConfig: .*/\\.?tsconfig.*\\.(js|json)$")
                 .regex(".*/\\.?tsconfig.*\\.(js|json)$"),
             ofFile("jsconfig", "/extra-icons/jsconfig.svg", "JSConfig: .*/\\.?jsconfig.*\\.(js|json)$")
@@ -365,8 +367,6 @@ public class ExtraIconProvider extends BaseIconProvider implements DumbAware {
                 .regex(".*/io/mockk/settings\\.properties"),
             ofFile("storybook", "/extra-icons/storybook.svg", "Storybook: *.stor(y|ies).(js|jsx|ts|tsx|mdx)")
                 .regex(".*\\.stor(y|ies)\\.(js|jsx|ts|tsx|mdx)$"),
-            ofFile("eslint", "/extra-icons/eslint.svg", "Eslint files")
-                .regex(".*\\.eslintrc(\\.[cm]?js(on)?|\\.y(a)?ml)?$"),
 
             //
             // file plus its containing folder
@@ -817,7 +817,7 @@ public class ExtraIconProvider extends BaseIconProvider implements DumbAware {
                 .eq("version", ".version", "versions", ".versions").end(".xml"),
             ofFile("vite", "/extra-icons/vite.svg", "Vite: vite.config.js")
                 .eq("vite.config.js"),
-            ofFile("vite_alt", "/extra-icons/vite_alt.svg", "Vite: vite.config.js")
+            ofFile("vite_alt", "/extra-icons/vite_alt.svg", "Vite: vite.config.js (alternative)")
                 .eq("vite.config.js"),
             ofFile("weblate", "/extra-icons/weblate.svg", "Weblate: .weblate")
                 .eq(".weblate"),
@@ -977,7 +977,7 @@ public class ExtraIconProvider extends BaseIconProvider implements DumbAware {
                 .end(".styl"),
             ofFile("ext_svg", "/extra-icons/svg.svg", "SVG: *.svg")
                 .end(".svg"),
-            ofFile("ext_svg", "/extra-icons/svg_alt.svg", "SVG: *.svg")
+            ofFile("ext_svg_alt", "/extra-icons/svg_alt.svg", "SVG: *.svg (alternative)")
                 .end(".svg"),
             ofFile("ext_tf", "/extra-icons/terraform.svg", "Terraform: *.hcl, *.tf, *.tf.json")
                 .end(".hcl", ".tf", ".tf.json"),